Refactor ShaderPropertySet to use uint32 chunks, updated Shader::GetRevisionNumber() to return int (so json values dont overflow)

+++ src/rendering/Shader.hpp
@@ -85,7 +85,7 @@ public:
 
     /// ===== Serialization only =====
     HYP_METHOD(Property = "RevisionNumber", NoScriptBindings)
-    uint64 GetRevisionNumber() const;
+    int GetRevisionNumber() const;
     /// ==============================
 
     Shader() = default;
